[VertexAI] Remove redundant tests (#6762)
There is an extra copy of some tests inside ` test/.../common`
directory. This change removes them, and in the process:

- Adds missing tests to the correct test file
- Gets rid of duplicated tests and test-util code
- Adds an exception type for quota exceeded
