+++ firebase-vertexai/gradle.properties
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-version=16.2.1
+version=16.3.0
 latestReleasedVersion=16.2.0
