+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/Exceptions.kt
@@ -59,6 +59,8 @@ internal constructor(message: String, cause: Throwable? = null) : RuntimeExcepti
               UnknownException(cause.message ?: "", cause.cause)
             is com.google.firebase.vertexai.common.ContentBlockedException ->
               ContentBlockedException(cause.message ?: "", cause.cause)
+            is com.google.firebase.vertexai.common.QuotaExceededException ->
+              QuotaExceededException(cause.message ?: "", cause.cause)
             else -> UnknownException(cause.message ?: "", cause)
           }
         is TimeoutCancellationException ->
@@ -162,6 +164,14 @@ internal constructor(location: String, cause: Throwable? = null) :
  * [Firebase documentation.](https://firebase.google.com/docs/vertex-ai/faq-and-troubleshooting#required-apis)
  */
 public class ServiceDisabledException
+internal constructor(message: String, cause: Throwable? = null) :
+  FirebaseVertexAIException(message, cause)
+
+/**
+ * The request has hit a quota limit. Learn more about quotas in the
+ * [Firebase documentation.](https://firebase.google.com/docs/vertex-ai/quotas)
+ */
+public class QuotaExceededException
 internal constructor(message: String, cause: Throwable? = null) :
   FirebaseVertexAIException(message, cause)
